changed for loops to have ints instead of lls

+++ header.cpp
@@ -14,11 +14,12 @@ using indexed_set = tree<TI,null_type,less<TI>,rb_tree_tag,tree_order_statistics
 #define all(x) (x).begin(), (x).end()
 #define rall(x) (x).rbegin(), (x).rend()
 #define uni(x) (x).erase(unique(all(x)), (x).end())
-#define rep(i, n) for (ll i = 0; i < (ll)(n); ++i)
-#define rep1(i, n) for (ll i = 1; i <= (ll)(n); ++i)
+#define rep(i, n) for (int32_t i = 0; i < (int32_t)(n); ++i)
+#define rep1(i, n) for (int32_t i = 1; i <= (int32_t)(n); ++i)
 const ld pi = 4.0*atanl(1.0);
 const ll infll = (ll)(1e18) + 10;
 const ll mod = powl(10, 9) + 7;
+
 
 int main() {ios::sync_with_stdio(false);cin.tie(0);cout.tie(0);
     //freopen("e.in", "r", stdin);
@@ -29,6 +30,36 @@ int main() {ios::sync_with_stdio(false);cin.tie(0);cout.tie(0);
     for(ll &x : arr) cin >> x;
     return 0;
 }
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
